Fixes psalm issues and composer dependencies
