+++ composer.json
@@ -42,20 +42,20 @@
         "nette/php-generator": "^4.0.1",
         "nikic/php-parser": "^4.13.2",
         "psr/container": "^1.1|^2.0",
+        "psr/event-dispatcher": "^1.0",
         "psr/http-factory": "^1.0",
+        "psr/http-factory-implementation": "^1.0",
         "psr/http-message": "^1.0",
         "psr/http-server-middleware": "^1.0",
         "psr/log": "1 - 3",
         "psr/simple-cache": "2 - 3",
         "spiral/attributes": "^2.8|^3.0",
         "spiral/composer-publish-plugin": "^1.0",
-        "symfony/console": "^5.3.7|^6.0",
+        "symfony/console": "^6.0",
         "symfony/finder": "^5.3.7|^6.0",
         "symfony/mailer": "^5.1|^6.0",
         "symfony/translation": "^5.1|^6.0",
-        "vlucas/phpdotenv": "^5.4",
-        "psr/http-factory-implementation": "^1.0",
-        "psr/event-dispatcher": "^1.0"
+        "vlucas/phpdotenv": "^5.4"
     },
     "autoload": {
         "files": [
