+++ src/Bridge/Dotenv/composer.json
@@ -31,8 +31,9 @@
         "vlucas/phpdotenv": "^5.4"
     },
     "require-dev": {
-        "spiral/boot": "^3.0",
+        "mockery/mockery": "^1.5",
         "phpunit/phpunit": "^9.5.20",
+        "spiral/boot": "^3.0",
         "vimeo/psalm": "^4.27"
     },
     "autoload": {
