+++ src/Broadcasting/psalm.xml
@@ -15,4 +15,11 @@
     <projectFiles>
         <directory name="src" />
     </projectFiles>
+    <issueHandlers>
+        <UndefinedClass>
+            <errorLevel type="suppress">
+                <referencedClass name="Spiral\Bootloader\Http\HttpBootloader" />
+            </errorLevel>
+        </UndefinedClass>
+    </issueHandlers>
 </psalm>
