+++ src/Console/composer.json
@@ -29,10 +29,10 @@
     "require": {
         "php": ">=8.1",
         "spiral/core": "^3.0",
+        "spiral/events": "^3.0",
         "spiral/hmvc": "^3.0",
         "spiral/tokenizer": "^3.0",
-        "symfony/console": "^5.3.7|^6.0",
-        "spiral/events": "^3.0"
+        "symfony/console": "^6.0"
     },
     "require-dev": {
         "spiral/boot": "^3.0",
