+++ src/Filters/composer.json
@@ -28,17 +28,18 @@
     ],
     "require": {
         "php": ">=8.1",
-        "spiral/auth": "^3.0",
+        "psr/http-message": "^1.0",
         "spiral/attributes": "^2.8|^3.0",
+        "spiral/auth": "^3.0",
         "spiral/core": "^3.0",
-        "spiral/models": "^3.0",
-        "spiral/validation": "^3.0",
         "spiral/hmvc": "^3.0",
-        "psr/http-message": "^1.0"
+        "spiral/models": "^3.0",
+        "spiral/validation": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5.20",
         "mockery/mockery": "^1.5",
+        "nyholm/psr7": "^1.5",
+        "phpunit/phpunit": "^9.5.20",
         "vimeo/psalm": "^4.27"
     },
     "autoload": {
