+++ src/Scaffolder/composer.json
@@ -40,12 +40,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5.20",
-        "spiral/boot": "^3.0",
-        "spiral/console": "^3.0",
-        "spiral/core": "^3.0",
-        "spiral/http": "^3.0",
-        "spiral/prototype": "^3.0",
-        "spiral/queue": "^3.0",
+        "spiral/framework": "^3.0",
         "vimeo/psalm": "^4.27"
     },
     "autoload": {
