+++ src/Snapshots/composer.json
@@ -27,7 +27,10 @@
         }
     ],
     "require": {
-        "php": ">=8.1"
+        "php": ">=8.1",
+        "spiral/exceptions": "^3.0",
+        "spiral/files": "^3.0",
+        "symfony/finder": "^5.3.7|^6.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5.20",
