+++ src/Stempler/composer.json
@@ -27,9 +27,10 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
         "ext-json": "*",
-        "myclabs/deep-copy": "^1.9"
+        "php": ">=8.1",
+        "myclabs/deep-copy": "^1.9",
+        "spiral/core": "^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5.20",
